Move configtarget conversion pkg under pkg/util/vsphere
populateConfigTargetDevices/PopulateStatus are pure govmomi-type
conversion helpers consumed only by controllers/configtarget; nothing
in pkg/providers/vsphere imports them. Relocate the package to
pkg/util/vsphere/configtarget, alongside the other pure conversion/
helper packages already there (client, datastore, library, vm),
per review feedback on PR #1732.
