+++ pkg/util/vsphere/configtarget/convert.go
