+++ pkg/util/vsphere/configtarget/status.go
